fix: now verify token function checks also email

+++ src/lib/server/auth/auth-utils.ts
@@ -5,34 +5,36 @@ import type { Cookies } from "@sveltejs/kit";
 import { SESSION_ID_LEN } from "$configs/fields-length";
 
 import type { Database } from "../db";
-import { createToken, deleteAllTokensByUserId, type DbToken, type TOKEN_TYPE, getTokenByUserId, deleteToken } from "../db/tokens";
+import { createToken, deleteAllTokensByUserId, type DbToken, TOKEN_TYPE, getTokenByUserId, deleteToken } from "../db/tokens";
 
-export async function generateToken(db: Database, userId: string, type: TOKEN_TYPE): Promise<DbToken | undefined> {
+export async function generateToken(db: Database, userId: string, email: string, type: TOKEN_TYPE): Promise<DbToken | undefined> {
+  if (!userId || !email || !type) return;
+
   await deleteAllTokensByUserId(db, userId, type);
 
-  const token = await createToken(db, { userId, type });
+  const token = await createToken(db, { userId, type, email });
   if (!token) return;
 
   return token;
 }
 
-export async function verifyToken(db: Database, userId: string, token: string, type: TOKEN_TYPE): Promise<boolean> {
+export async function verifyToken(db: Database, userId: string, token: string, type: TOKEN_TYPE, email?: string): Promise<string | undefined> {
+  if (!userId || !token || !type) return;
+
   const tokenFromDatabase = await getTokenByUserId(db, userId, type);
-  if (!tokenFromDatabase || tokenFromDatabase.token !== token) {
-    return false;
-  }
+  if (!tokenFromDatabase || tokenFromDatabase.token !== token) return;
 
   const deletedToken = await deleteToken(db, token, type);
-  if (!deletedToken) {
-    return false;
-  }
+  if (!deletedToken) return;
 
   const isExpired = !isWithinExpirationDate(tokenFromDatabase.expiresAt);
-  if (isExpired) {
-    return false;
-  }
+  if (isExpired) return;
 
-  return true;
+  // in case of verify email and reset password journey, we also need to verify the email
+  if (email && type !== TOKEN_TYPE.EMAIL_CHANGE && email !== tokenFromDatabase.email) return;
+
+  // we return email because we need it in email change journey
+  return tokenFromDatabase.email;
 }
 
 export function setNewSession(lucia: Lucia, sessionId: string, cookies: Cookies) {
